Added size checking for transmutes.
Added "debug_checks" feature,
which enables size checking of transmuted values/pointers when it's enabled.

Added tests for the `peel_ref`, `wrap_ref`, `transmute_ref` macros.

Added tests ensuring that when the `"debug_checks"` feature is enabled,
everything panics when it should.

Documented that infer doesn't work with `peel_ref` or `wrap_ref`.

Changed the language from "transmute" to "cast" in the docs of safe functions.