Removed TransmutableInto
Removed it because it was a leaky abstraction,
by allowing transmutes by value and by reference it was overly restrictive.
Only allowing by-value transmutes wouldn't be too useful either.
Eh, might add it back at some point, if there's a clearly better way to do it.

Made some changes in `constmuck_internal` to accomodate this change.